Refactor favorites handling: enhance AddFavorite function to accept user ID and asset details, and implement data loading for favorites
